int to float changes

+++ data types/command argv.py
@@ -13,9 +13,9 @@ def div():
     div = num1 / num2
     return div
 
-num1 = int(sys.argv[1])
+num1 = float(sys.argv[1])
 operation = sys.argv[2]
-num2 = int(sys.argv[3])
+num2 = float(sys.argv[3])
 
 if operation == "add":
     output = add()
